Moved FileManager to utils. Fixed imports in project
