+++ src/utils/include/AODPProject/utils/FileData.h
@@ -1,0 +1,16 @@
+#pragma once
+#include <string>
+#include <vector>
+
+
+struct FileData {
+	std::string fileName;
+	int numberOfIterations;
+	int colonySize;
+};
+
+class FileManager {
+public:
+	void loadFromFile(std::vector<std::vector<int> >& edges, int& numberOfVertexes, std::string fileName);
+    std::vector<FileData> loadTestInitFile(std::string initFileName);
+};
